+++ src/compress/lz10.zig
@@ -72,12 +72,12 @@ pub fn blockEncoding(block: lz.Block) u1 {
 pub const Compress = lz.Compress(lz10);
 pub const Decompress = lz.Decompress(lz10);
 
-// TODO: Tests
-
 comptime {
     _ = Compress;
     _ = Decompress;
 }
+
+// TODO: Fuzzing in 0.17 as it is broken in 0.16
 
 const testing = std.testing;