+++ src/compress/yaz.zig
@@ -90,12 +90,12 @@ pub fn blockEncoding(block: lz.Block) u1 {
 pub const Compress = lz.Compress(yaz);
 pub const Decompress = lz.Decompress(yaz);
 
-// TODO: Tests
-
 comptime {
     _ = Compress;
     _ = Decompress;
 }
+
+// TODO: Fuzzing in 0.17 as it is broken in 0.16
 
 const testing = std.testing;